\name{dataPub}
\docType{data}
\alias{dataPub}
\title{Publishing productivity data}
\description{
Data on publishing productivity among academics.
}
\usage{data(PubProd)}
\format{
  \code{dataPub} contains the following objects:
  \describe{
    \item{\code{Corr}}{
      Matrix (7x7) with the correlation matrix of the variables. 
    }
    \item{\code{nobs}}{
      Scalar with the number of observations.
    }
  }
}
\source{Spirtes, P., Glymour, C., and Scheines, R. (2000). Causation, prediction and search (2nd
edition). \emph{Cambridge, MA: The MIT Press}. pages 1-16.
}
\references{
Drton, M. and Perlman, M. D. (2008). A SINful approach to Gaussian graphical model
selection. \emph{J. Statist. Plann. Inference} 138, 1179-1200.

D. Altomare, G. Consonni and L. La Rocca (2012). Objective Bayesian search of Gaussian directed acyclic graphical models for ordered variables with non-local priors. \emph{Article submitted to Biometric Methodology}.
}
\keyword{datasets}
